Preview: Create a single instance of the preview panel
Every time the user pushed the UP/DOWN key a new instance of the preview
panel was created, resulting in bad performance. Additionally, the
preview panel was only created after the first key push, giving the
feeling to the user that it's even slower.

Cleanup the code in order to make sense and changed to create a single
instance. By using the UP/DOWN key it only updates the already existing
preview panel. Much smoother result and much higher quality code :)