Preview: Fix inconsistent rulers behavior
Cleanup the code and use maths instead of multiple if statements
to make the rulers behave the same for all preview images.

Additionally, clear the preview when node does not support
preview and user can zoom in and out using Ctrl + mouse wheel.

The zoom value is now a fixed label showing the zoom percentage.